Since "The One Major Rule" ( http://www.minecraft.net/terms ) is "do not distribute", don't include the official minecraft server binary.
Since I don't want to add an uneccesary extra step to the
installation, I have added in a craftbukkit build instead. Since
craftbukkit is under LGPL, this should be OK. In theory I should
include a copy of the LGPL and the craftbukkit source code when
distributing the craftbukkit binary, but since those are so easily
available online I will just hope that nobody is pedantic enough to
care. If you want either, write me and I will send them to you :P.